Update GetmRNAorCDSforProtein.py
Commented out the return statement for function `mine_the_cds_sequence_from_nuccore_seq` where it tries to mine the CDS from features because third approach extracting a CDS from the coded_by information is more reliable until I have a step where validity of sequence mined can be determined and that way can return None if not valid or the sequence if it is. It was working that if it returned anything it assumed it was valid but that may be bad.