mention I plan to make a file no needing gtf
mention I plan to make a file no needing gtf file for conversion. It would be more broadly applicable to just use YeastMine and that way any file will work. I currently have a file that makes a list that includes the standard (common) names if provided a list of systematic names, but the current form doesn't keep the order in which the original list is provided. Keeping the order will make subsequent steps easier.