change to `is None` which is better
change to `is None`, which is better according to  PEP 8 Style Guide, the preferred way to compare something to None are the patterns if cond is None or if cond is not None.

The statement is is used for object identity, it checks if objects refer to the same instance. Since None is a singleton this statement can be used to safely check whether a variable points to None.

The == and != statement refers to equality, which means they call the __eq__ method of the referred objects. But this function can differ depending on the implementation of the objects.