change output file from append to write
When developing I realized while "append" sounded good because going to be writing many lines as a loop runs, it caused multiple copies to be appended to same file as I was testing. Probably want "write" and still shouldn't close until it totally exits that look since nested. So shouldn't make less efficient.