revers to using pd.read_csv()
I had been debugging something else it turns out and had thought pd.read_table() was needed because when changed that it was around the time things worked again. Questioned it though the documentation made it look like StringIO objects worked with both. Turns out it was not needed and pasting the code into the cell and not importing had fixed because using caps and underscore both in script name and function name seemed to cause problem, despite that caps at beginning of UCSC to karyotype script was just fine?!!!1