remove exception handling for string vs file
Remove exception handling for string vs file. Turns out Pandas pd.read_csv() can handle both out-of-the-box, and so I don't need acrobatics of sorting which it is and then using StringIO on the string before passing to pd.read_csv(). (I had thought this was the case at the start but when I was having import issues, apparently because of capitalized name of file file and function, I had convinced myself erroneously while attempt debugging that `pd.read_csv()` couldn't take a string.