convert first two parts to using Biopython for super efficiency boost
convert parsing alignment and extracting sequences and coordinates to Biopython because doing with strings and list was super slow when the sequence was greater than 10K. And got worse with size of sequence which made scaling to a lot of regions in a large sequence take tens of minutes and scaling up to do that in multiple sequences add up to hours. I was already doing the saving output files with Biopython and so that part didn't need much editing.