Close #1 - make work with chromosome subsets again and bonus outcome was fixing order along X-axis
Making it work again with chromosomes or subsets of chromosomes alone, seems to have also addressed issues with order. Seems applying categorical to chromosomes but not the other columns in the data allowed ordering to match the GTF as a result. That was totally serendipitous because I had with all columns that were strings as `category` and it put human chromosome X immediately after seven (and eleven in front of ten) and the yeast roman numerals were jumbled. It got horrible as far as order of chromosomes when I made all `object`, which most strings should be, but when I made `seqname` categoery and all other string columns as `object` then the human and the yeast were both ordered matching the GTF. Which is what I had wanted all along. Fixing to work with chromosome subsets required adding `observed=True` to the groupby steps so that not all categoricals used, resulting in just those with rows. So I addressed the original and resulting later-indicated issues in https://github.com/fomightez/sequencework/issues/1 .