add increasing bitscore when see a case where start and end seem on different strands
had 11 cases where start and end seem on different strands without adding this bump up in those cases. four out of four tested were fixed with this fix. None of four had omega. (Actually I ended up hand checking and none of 11 had the omega intron. However, I worry this approach may be slightly flawed though in that I didn't propogate over that increase in bitscore cutoff where checked against full genomic with the intron, and thus a shift to less hits with the full genomic could be influenced AND MISSED by this and I just got lucky that none of those involved had omega intron? By the way it isn't easy to proppgate since the checking for omega intron is actually a separate script so I can't just re-adjust setting of bitscore cutoff in one place but would need to pass it.)