add way to pass an adjusted bitscore
add way to pass an adjusted bitscore so that if have to adjust and passing a dataframe for the first round of BLAST, still using same score as cutoff when checking for intron presence. Was because of change in `find_mito_fungal_lsu_rRNA_and_check_for_omega_intron.py` to deal with a few apparent edge cases.