restrict to using ids that match the 21S sequence without the intron
restrict to using ids that match to best match to the 21S sequence without the intron for considering intron present or not. I found when started trying against genome assembly. Poorer scoring sequences that spurious looked to match near intron insertion point were confusing the call of intron present or not.